test: split test_junctions.py into core/analyses/inversion files
